refactor: EmptyProjectList, ProjectList 별도 파일로 분리 및 ui 모듈로 이동
